+++ lewp/src/config/mod.rs
@@ -1,6 +1,9 @@
 //! Configuration data structures used by [lewp](crate).
 
-use {crate::fh::FileHierarchy, std::sync::Arc};
+use {
+    crate::{css::Register as CssRegister, fh::FileHierarchy, Lewp},
+    std::sync::Arc,
+};
 
 /// Various config values for adjusting the modules behavior.
 pub struct ModuleConfig {}
@@ -24,20 +27,45 @@ pub struct PageConfig {
     pub viewport_tag: bool,
     /// The file hierarchy where the page is working on.
     pub fh: Option<Arc<FileHierarchy>>,
+    /// The CSS register containing all required CSS data of modules and pages.
+    pub css_register: Option<Arc<CssRegister>>,
 }
 
 impl PageConfig {
     /// Creates a new instance with default values.
-    pub fn new(fh: Option<Arc<FileHierarchy>>) -> Self {
+    pub fn new(
+        viewport_tag: bool,
+        fh: Option<Arc<FileHierarchy>>,
+        css_register: Option<Arc<CssRegister>>,
+    ) -> Self {
+        Self {
+            viewport_tag,
+            fh,
+            css_register,
+        }
+    }
+}
+
+impl From<&Lewp> for PageConfig {
+    fn from(lewp: &Lewp) -> Self {
+        let fh = match &lewp.fh {
+            Some(f) => Some(f.clone()),
+            None => None,
+        };
+        let css_register = match &lewp.css_register {
+            Some(c) => Some(c.clone()),
+            None => None,
+        };
         Self {
             viewport_tag: true,
             fh,
+            css_register,
         }
     }
 }
 
 impl Default for PageConfig {
     fn default() -> Self {
-        Self::new(None)
+        Self::new(true, None, None)
     }
 }