+++ lewp/tests/loop-detection.rs
@@ -135,7 +135,7 @@ fn loop_detection() {
     let module = modules::Header::new();
     let mut page = HelloWorldPage {
         modules: vec![],
-        config: PageConfig::new(None),
+        config: PageConfig::default(),
     };
     page.add_module(module.into_module_ptr());
     page.build();