+++ lewp/tests/sub-module.rs
@@ -167,7 +167,7 @@ fn submodule() {
     let module = modules::Header::new();
     let mut page = HelloWorldPage {
         modules: vec![],
-        config: PageConfig::new(None),
+        config: PageConfig::default(),
     };
     page.add_module(module.into_module_ptr());
     let dom = page.build();